Extract ApplicationScraper class.
